\docType{package}
\name{gte-package}
\alias{gte-package}
\title{Generalized Turnbull's Estimator}
\description{
  The \code{gte} function computes the generalized
  Turnbull's estimator proposed by Dehghan and Duchesne
  (2011). It is a nonparametric estimator of a conditional
  survival function given a vector of continuous covariates
  that can handle interval-censored lifetimes.
}
\details{
  \tabular{ll}{ Package: \tab gte\cr Type: \tab Package\cr
  Version: \tab 1.2-1\cr Date: \tab 2014-01-03\cr License:
  \tab GPL-2\cr }
}
\author{
  Mohammad Hossein Dehghan, Thierry Duchesne and Sophie
  Baillargeon

  Maintainer: Thierry Duchesne
  <thierry.duchesne@mat.ulaval.ca>
}
\references{
  Dehghan, M. H. and Duchesne, T. (2011). A generalization
  of Turnbull's estimator for nonparametric estimation of
  the conditional survival function with interval-censored
  data. \emph{Lifetime Data Analysis}, \bold{17}, 234-255.
}
\keyword{package}

